backend/pyproject.toml
[project]
name = "anime-rec"
version = "0.1.0"
description = "Anime recommendation website backend (FastAPI + Postgres + Redis)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlmodel>=0.0.21",
  "psycopg-binary>=2.9.11",
  "pydantic-settings>=2.4.0",
  "redis>=5.0.7",
  "httpx>=0.27.0",
  "numpy>=2.0.0",
  "python-dotenv>=1.0.1",
  "dotenv>=0.9.9",
  "psycopg2-binary>=2.9.11",
  "scikit-learn>=1.7.2",
  "python-jose>=3.5.0",
  "passlib[bcrypt]==1.7.4",
  "bcrypt==3.2.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
